==> bin/1.1-venn-0.05.Rmd <==
---
title: "Venn Diagrams of genes that are differentially regulated by anc-1"
author: "Amir Levine"
date: "2018-12-13"
output: md_document
---

Load Packages
```{r packages}
#source("https://bioconductor.org/biocLite.R")
# install.packages("VennDiagram") ; library(VennDiagram)
if (!require(eulerr)) {BiocManager::install("eulerr"); library(eulerr)} # Plotting Euler Diagrams
```

Set experiment path
```{r set-path}
path = "experiments/2018-04-24-venn/"
```

## Top regulated genes
Get top 15 genes per overlap by their adjusted *p*-values.
```{r top-15}
reg <- NT$Gene_name[NT$Gene_stable_ID %in% intersect(DEG$AM.up,DEG$N2.up)][1:15]
write(reg, paste0(path,"results/topGenes-up.txt"))
reg <- NT$Gene_name[NT$Gene_stable_ID %in% intersect(DEG$AM.dn,DEG$N2.dn)][1:15]
write(reg, paste0(path,"results/topGenes-dn.txt"))
```

## Genes that passed a significance threshold of adjusted *p*-value < 0.05

Load Processed Data
```{r load-data}
load("data/2018-11-27-geneExpression-pval-thresh-0.05.rdata")
```

## **Hypergeometric distribution test** to check if gene overlap is significant
Check the *p*-values of overlaps that are equal, or greater than the measured overlap.

The total population are genes that appeard in the results of either AM140 or N2 CEL-seq, as these are the measured genes.
```{r background-pop-size}
totalpop <- length( 
  union(
    union(DEG$N2.up,DEG$N2.dn),
    union(DEG$AM.up,DEG$AM.dn)
  ) 
)
```

### Upregulated genes
```{r upregulated-size}
# Genes that were upregulated in AM140 animals
sample1 <- length(DEG$AM.up)
# Genes that were upregulated in N2 animals
sample2 <- length(DEG$N2.up)
# Genes that were upregulated in both AM140 and N2 animals
overlap <- length( intersect(DEG$AM.up,DEG$N2.up) )
```

Hypergeometric test
```{r test-overlap-upreg}
hyper.up <- sum( dhyper(x = overlap:sample1,
                        m = sample2,
                        n = totalpop - sample2,
                        k = sample1) )
print(hyper.up)
```

### Downregulated genes
```{r downregulated-size}
# Genes that were upregulated in AM140 animals
sample1 <- length(DEG$AM.dn)
# Genes that were upregulated in N2 animals
sample2 <- length(DEG$N2.dn)
# Genes that were upregulated in both AM140 and N2 animals
overlap <- length( intersect(DEG$AM.dn,DEG$N2.dn) )
```

Hypergeometric test
```{r test-overlap-downreg}
hyper.dn <- sum( dhyper(x = overlap:sample1,
                        m = sample2,
                        n = totalpop - sample2,
                        k = sample1) )
print(hyper.dn)
```

## Plot Venn Diagrams

Upregulated genes
```{r euler-plot-up}
# Gene counts per category
list = setNames(c(
  length(setdiff(DEG$N2.up,DEG$AM.up)),
  length(setdiff(DEG$AM.up,DEG$N2.up)),
  length(intersect(DEG$AM.up,DEG$N2.up))
), c("N2","AM140","N2&AM140"))

vd <- euler(list, shape = "ellipse") # Generate Euler

cairo_pdf(
  filename = paste0(path,"results/2019-02-09-euler-anc1-up-0.05.pdf"),
  width = 1, height = 1)

plot(
  vd, 
  lwd = 1,
  fills = list(fill = c("indianred2","royalblue2")),
  quantities = list(fontsize = 7),
  labels = F
)

dev.off()
```

Downregulated genes
```{r euler-plot-dn}
# Gene counts per category
list = setNames(c(
  length(setdiff(DEG$N2.dn,DEG$AM.dn)),
  length(setdiff(DEG$AM.dn,DEG$N2.dn)),
  length(intersect(DEG$AM.dn,DEG$N2.dn))
), c("N2","AM140","N2&AM140"))

vd <- euler(list, shape = "ellipse") # Generate Euler

cairo_pdf(
  filename = paste0(path,"results/2019-02-09-euler-anc1-dn-0.05.pdf"),
  width = 1, height = 1)

plot(
  vd, 
  lwd = 1,
  fills = list(fill = c("indianred2","royalblue2")),
  quantities = list(fontsize = 7),
  labels = F
)

dev.off()
```
