==> bin/2.7-Tbox-enrich-heatmap.Rmd <==
---
title: "T-box transcription factors heatmap"
author: "Amir Levine"
date: "2018-12-31"
output: md_document
---

## Load packages and datasets.
```{r packages}
if (!require(tidyverse)) {install.packages("tidyverse"); library(tidyverse)} # Data pre-processing
if (!require(pheatmap)) {biocLite("pheatmap"); library("pheatmap")} # Plot heatmap
```

PWMEnrich object of Tbox transcrption factors motifs enrichement in the promoters of *anc-1* RNAi regulateed genes. 
Note that the `PWMEnrich object` refers to anc-1 RNAi regulated genes set by a p < 0.1 threshold, in order to get a broader coverage of possible regulated genes.
```{r datasets}
load("data/Celegans-promoters.rdata",verbose = TRUE)
#load("experiments/2018-07-11-tbx-motifs/data/2018-10-25-Tbox-anc1-reg-enriched-pval.1.Rdata",verbose = TRUE)
load("experiments/2018-07-11-tbx-motifs/data/2019-02-07-Tbox-anc1-reg-enriched.Rdata",verbose = TRUE)
```

## Data Pre-Processing
Focus on the genes that are down-regulated by *anc-1*, as those are enriched for T-box motifs.
```{r pval-matrix}
# Create p-values matrix. Obtain the *p*-values of motifs per sequence. Rows correspond to the different input sequences and the columns correspond to motifs. `sequence.nobg` contains the raw affinity scores.`sequence.bg` contains the corresponding P-values.
tbx.pv.dn.mat <- res.dn$sequence.bg

# Get the "human-readable" gene names and assign as the matrix rownames.
gnames <- mRNA.seqs.promoter$external_gene_name[which(mRNA.seqs.promoter$ensembl_gene_id %in% rownames(tbx.pv.dn.mat))]
rownames(tbx.pv.dn.mat) <- gnames

# Rename the column names to be the T-box protein names.
colnames(tbx.pv.dn.mat) <- c("TBX-33","TBX-38","TBX-39","TBX-40","TBX-43")
```

# Convert the TBX binding-sites p-values matrix to binary
```{r tbx-matrix}
# Change p-value to boolean. TRUE is statistically significant (*p*-value < 0.01).
tbx.bool.dn.mat <- (tbx.pv.dn.mat < .01)

# Index of amount of T-box transcription factors binding motif
TBXTFBM <- rowSums(tbx.bool.dn.mat*1) != 0

# Filter genes w/o T-box transcription factors binding motifs from the p-values matrix
tbx.pv.dn.mat <- tbx.pv.dn.mat[TBXTFBM,]

# Filter genes w/o T-box transcription factors binding motifs from the Boolean matrix
tbx.bool.dn.mat <- tbx.bool.dn.mat[TBXTFBM,]
```

```{r}
### For new version only, can delete if irrelevant

tbx.pv.dn.mat <- apply(tbx.pv.dn.mat, c(1,2), FUN = function(x) {
  if (x<.01) {
    x <- 0
  } else if (x<.05) {
    x <- .5
  } else {
    x <- 1
  }
}
)

tbx.pv.dn.mat <- tbx.pv.dn.mat[TBXTFBM,]
```

## Plotting Heatmaps

Setting the path for saving the heatmaps
```{r}
path = "experiments/2018-12-30-tbx-heatmap/results/"
```

### Plotting the p<0.01 from the PWMEnrich of p<0.05 by anc-1 RNAi
Annotations for matrix of binding sites presence
```{r}
rownames(tbx.bool.dn.mat)

tbtf = "T-box transcription factor"
tf = "Transcription factor"
fb = "F-box protein-related"
pr = "Proteostasis-related"


Function <- c(
  NA,tf,NA,tf,NA,NA,NA,pr,tbtf,tf,
  tf,tf,NA,fb,NA,NA,NA,fb,NA,NA,
  fb,NA,pr,NA,fb,NA,NA,NA,NA,NA,
  NA,tbtf,NA,NA
)

ann <- data.frame("Function" = Function
                  #"Fold-Change (WT)" = FC.N2
                  #"Fold-Change (polyQ35)" = FC.AM
)
rownames(ann) <- rownames(tbx.bool.dn.mat)

# Fold-change levels of the selected genes

```

Heatmap of T-box transcription factors binding sites binary presence
```{r}
### Old Version - no p< .05

row_dend <- hclust(dist(tbx.bool.dn.mat, method = "binary"),
                   method = "complete") # row clustering
col_dend <- hclust(dist(t(tbx.bool.dn.mat), method = "binary"),
                   method = "complete") # column clustering

pheatmap(
  tbx.pv.dn.mat*1,
  
  # Colors of the cells
  color = c("white","black"),
  border_color = "lightgrey",
  
  # dimensions of the cells
  cellwidth = 7,
  cellheight = 7,
  
  # Clustering
  cluster_rows = row_dend,
  cluster_cols = col_dend,
  
  # dendogram graphics
  treeheight_row = 5,
  treeheight_col = 5,
  cutree_rows = 8,
  
  # Remove the legend
  legend = FALSE,
  
  # Fonts
  fontsize = 7,
  
  # Gene annotations
  annotation_row = ann,
  drop_levels = TRUE,
  
  # Export figure
  filename = paste0(path,"2019-02-12-hm",".pdf")
)

### New Version - w/ p< .05

row_dend <- hclust(dist(tbx.bool.dn.mat, method = "binary"),
                   method = "complete") # row clustering
col_dend <- hclust(dist(t(tbx.bool.dn.mat), method = "binary"),
                   method = "complete") # column clustering

pheatmap(
  tbx.pv.dn.mat,
  
  # Colors of the cells
  color = c("black","darkgrey","white"),
  border_color = "lightgrey",
  
  # dimensions of the cells
  cellwidth = 7,
  cellheight = 7,
  
  # Clustering
  cluster_rows = row_dend,
  cluster_cols = col_dend,
  
  # dendogram graphics
  treeheight_row = 5,
  treeheight_col = 5,
  cutree_rows = 8,
  
  # Remove the legend
  legend = FALSE,
  
  # Fonts
  fontsize = 7,
  
  # Gene annotations
  annotation_row = ann,
  drop_levels = TRUE,
  
  # Export figure
  filename = paste0(path,"2019-02-12-w.05-hm",".pdf")
)
```


### Plotting the p<0.01 from the PWMEnrich of p<0.1 by anc-1 RNAi
Annotations for matrix of binding sites presence
```{r}
tbtf = "T-box transcription factor"
tf = "Transcription factor"
chrm = "Chromatin-related"
prts = "Proteostasis-related"
fb = "F-box protein-related"

Function <- c(tbtf, tf, chrm, prts, prts, tf, tf, tbtf, "", "",
              tf, "", "", "", "", "", "", tf, "", "",
              tf, fb, "", "", chrm, tf, "", "", fb, "", 
              tbtf, "", "", "", fb, "", "", "", "", fb,
              chrm, "", prts, tf, "", "", ""
)
# Decided not to include the expression changes
#FC.N2 <- AM.NT$log2FoldChange.N2[AM.NT$Gene_name %in% rownames(tbx.bool.dn.mat)]
#FC.AM <- AM.NT$log2FoldChange.AM140[AM.NT$Gene_name %in% rownames(tbx.bool.dn.mat)]

ann <- data.frame("Function" = Function
                  #"Fold-Change (WT)" = FC.N2
                  #"Fold-Change (polyQ35)" = FC.AM
)
rownames(ann) <- rownames(tbx.bool.dn.mat)

# Fold-change levels of the selected genes

```

Heatmap of T-box transcription factors binding sites binary presence
```{r}
row_dend <- hclust(dist(tbx.bool.dn.mat, method = "binary"),
                   method = "complete") # row clustering
col_dend <- hclust(dist(t(tbx.bool.dn.mat), method = "binary"),
                   method = "complete") # column clustering

pheatmap(tbx.bool.dn.mat*1,
         
         # Colors of the cells
         color = c("white","black"),
         border_color = "grey",
         
         # dimensions of the cells
         cellwidth = 7,
         cellheight = 7,
         
         # Clustering
         cluster_rows = row_dend,
         cluster_cols = col_dend,
         cutree_rows = 9,
         
         # dendogram graphics
         treeheight_row = 5,
         treeheight_col = 5,
         
         # Remove the legend
         legend = FALSE,
         
         # Fonts
         fontsize = 7,
         
         # Gene annotations
         annotation_row = ann,
         drop_levels = TRUE,

         # Export figure
         filename = paste0(path,"2018-12-31-hm",".pdf")
         )
```

```
> sessionInfo()
R version 3.5.1 (2018-07-02)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows >= 8 x64 (build 9200)

Matrix products: default

locale:
[1] LC_COLLATE=English_Israel.1252  LC_CTYPE=English_Israel.1252    LC_MONETARY=English_Israel.1252
[4] LC_NUMERIC=C                    LC_TIME=English_Israel.1252    

attached base packages:
 [1] stats4    parallel  grid      stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
 [1] biomaRt_2.38.0        BiocInstaller_1.32.1  pheatmap_1.0.10       gplots_3.0.1          RColorBrewer_1.1-2   
 [6] psych_1.8.10          corrplot_0.84         Hmisc_4.1-1           Formula_1.2-3         survival_2.42-3      
[11] lattice_0.20-35       PWMEnrich_4.18.0      Biostrings_2.50.0     XVector_0.22.0        IRanges_2.16.0       
[16] S4Vectors_0.20.0      BiocGenerics_0.28.0   bindrcpp_0.2.2        viridis_0.5.1         viridisLite_0.3.0    
[21] ComplexHeatmap_1.20.0 dendextend_1.9.0      forcats_0.3.0         stringr_1.3.1         dplyr_0.7.7          
[26] purrr_0.2.5           readr_1.1.1           tidyr_0.8.2           tibble_1.4.2          ggplot2_3.1.0        
[31] tidyverse_1.2.1      

loaded via a namespace (and not attached):
 [1] colorspace_1.3-2     rjson_0.2.20         class_7.3-14         modeltools_0.2-22    mclust_5.4.1        
 [6] evd_2.3-3            circlize_0.4.4       htmlTable_1.12       GlobalOptions_0.1.0  base64enc_0.1-3     
[11] rstudioapi_0.8       bit64_0.9-7          flexmix_2.3-14       AnnotationDbi_1.44.0 fansi_0.4.0         
[16] mvtnorm_1.0-8        lubridate_1.7.4      xml2_1.2.0           splines_3.5.1        mnormt_1.5-5        
[21] robustbase_0.93-3    knitr_1.20           jsonlite_1.5         seqLogo_1.48.0       broom_0.5.0         
[26] cluster_2.0.7-1      kernlab_0.9-27       BiocManager_1.30.3   compiler_3.5.1       httr_1.3.1          
[31] backports_1.1.2      assertthat_0.2.0     Matrix_1.2-14        lazyeval_0.2.1       cli_1.0.1           
[36] prettyunits_1.0.2    acepack_1.4.1        htmltools_0.3.6      tools_3.5.1          gtable_0.2.0        
[41] glue_1.3.0           Rcpp_0.12.19         Biobase_2.42.0       cellranger_1.1.0     trimcluster_0.1-2.1 
[46] gdata_2.18.0         nlme_3.1-137         fpc_2.1-11.1         rvest_0.3.2          gtools_3.8.1        
[51] XML_3.98-1.16        DEoptimR_1.0-8       zlibbioc_1.28.0      MASS_7.3-50          scales_1.0.0        
[56] hms_0.4.2            yaml_2.2.0           memoise_1.1.0        gridExtra_2.3        rpart_4.1-13        
[61] RSQLite_2.1.1        latticeExtra_0.6-28  stringi_1.2.4        checkmate_1.8.5      caTools_1.17.1.1    
[66] shape_1.4.4          rlang_0.3.0.1        pkgconfig_2.0.2      prabclus_2.2-6       bitops_1.0-6        
[71] bindr_0.1.1          htmlwidgets_1.3      bit_1.1-14           tidyselect_0.2.5     plyr_1.8.4          
[76] magrittr_1.5         R6_2.3.0             DBI_1.0.0            pillar_1.3.0         haven_1.1.2         
[81] whisker_0.3-2        foreign_0.8-70       withr_2.1.2          RCurl_1.95-4.11      nnet_7.3-12         
[86] modelr_0.1.2         crayon_1.3.4         KernSmooth_2.23-15   utf8_1.1.4           progress_1.2.0      
[91] GetoptLong_0.1.7     readxl_1.1.0         data.table_1.11.8    blob_1.1.1           digest_0.6.18       
[96] diptest_0.75-7       munsell_0.5.0       
```
