==> bin/3.8-TFs-indiv-reg-direction-enrichment.Rmd <==
---
title: "Check for an enrichment of C. elegans transcription factors motifs in the promoter of anc-1 targets"
output: html_document
---

## Load bioconductor and packages
```{r}
source("https://bioconductor.org/biocLite.R")
if (!require(tidyverse)) {install.packages("tidyverse"); library(tidyverse)}
if (!require(PWMEnrich)) {install.packages("PWMEnrich"); library(PWMEnrich)}
install.packages("VennDiagram") ; library(VennDiagram)
```

# Gene-lists and Sequences
Load data from upstream analyses.
```{r}
# Load CEL-seq gene lists
load("data/geneExpression.rdata")
# Load promoters of C. elegans genes. Data generated by "Celegans-promoters.R"
load("data/Celegans-promoters.rdata")
# Load the background PWMs applied to all sequences object, which is generated by "TFs-background.Rmd".
load("experiments/2018-09-02-TFs-motifs/data/background-celegans-TF.rdata")
```

# Motif enrichment scanning
Perform transcription factors enrichment analysis to sequences of interest -- the promoter sequences of all the genes that are significantly (p<0.1) regulated by *anc-1*, in both strains. A previous version of this analysis focused on the intersect of the two strains, with upregulated and downregulated genes analyzed separately.

Set all genes lists of interest.
```{r}
reg.genes <- list(N2.up = DEG$N2.up,
                  N2.dn = DEG$N2.dn,
                  AM.up = DEG$AM.up,
                  AM.dn = DEG$AM.dn)
```

Define function for motif enrichment analysis based on the bioconductor (PWMEnrich  package)[https://www.bioconductor.org/packages/release/bioc/html/PWMEnrich.html]
```{r}
me <- function(goi,mRNA.seqs.promoter,prs) {
  mRNA.seqs.promoter$CELseq.hits <- mRNA.seqs.promoter$ensembl_gene_id %in% goi
  return(motifEnrichment(prs[mRNA.seqs.promoter$CELseq.hits], bg.custom))
}
```

Apply function to each gene list and store results as a list file.
```{r}
useBigMemoryPWMEnrich(TRUE) # using large amount of memory
#registerCoresPWMEnrich(4) # Parallel execution Not supported in Windows

res <- lapply(reg.genes, me,mRNA.seqs.promoter,prs)

useBigMemoryPWMEnrich(FALSE) # disable using large amount of memory
#registerCoresPWMEnrich(NULL) # disable Parallel execution
```

It was a long run so save the generated object.
```{r}
save(res,file = "experiments/2018-09-02-TFs-motifs/data/TFs-motif-enrichment-results")
```

## Motif enrichment scanning to sampled random sequences as a control
Perform an analysis similar to the previous one, only with random sequences to serve as a negative control.

Check the regulated gene lists length to generate random lists, and generate lists.
```{r}
rand.len <- lapply(reg.genes,length)

rand.reg.genes <- list()

set.seed(sum(unlist(rand.len)))

for (trt in names(rand.len)) {
  rand.reg.genes[[trt]] <- sample(mRNA.seqs.promoter$ensembl_gene_id,
                                  rand.len[[trt]])
}
```

Apply motif enrichment analysis function to each gene list and store results as a list file.
```{r}
useBigMemoryPWMEnrich(TRUE) # using large amount of memory
#registerCoresPWMEnrich(4) # Parallel execution Not supported in Windows

rand.res <- lapply(rand.reg.genes,me,mRNA.seqs.promoter,prs)

useBigMemoryPWMEnrich(FALSE) # disable using large amount of memory
#registerCoresPWMEnrich(NULL) # disable Parallel execution
```

It was a long run so save the generated object.
```{r}
save(rand.res,file = "experiments/2018-09-02-TFs-motifs/data/TFs-motif-enrichment-results-random")
```
